+++ app/src/main/java/de/triology/universeadm/Paginations.java
@@ -35,20 +35,25 @@ import java.util.List;
  */
 public final class Paginations
 {
+  private static final int PAGING_DEFAULT_START = 0;
   private Paginations(){
     
   }
   
   public static <T> PagedResultList<T> createPaging(List<T> list, int start, int limit){
-    PagedResultList<T> result = null;
     int size = list.size();
-    if ( start < size ){
-      int end = start + limit;
-      if ( end > size ){
-        end = size;
-      }
-      result = new PagedResultList<>(list.subList(start, end), start, limit, size); 
+    int end = limit;
+    int actualStartValue = start;
+    if ( start < size ) {
+      end = start + limit;
+    } else {
+      actualStartValue = PAGING_DEFAULT_START;
     }
-    return result;
+
+    if ( end > size ) {
+      end = size;
+    }
+
+    return new PagedResultList<>(list.subList(actualStartValue, end), actualStartValue, limit, size);
   }
 }