+++ app/src/main/ui/src/hooks/useCasUser.tsx
@@ -1,0 +1,9 @@
+import { CasUserService} from "../services/CasUser";
+import { useAPI} from "./useAPI";
+import type {CasUser} from "../services/CasUser";
+
+export const useCasUser = () => {
+    const {data} = useAPI<CasUser>(CasUserService.get);
+    const user = data ?? {principal: "default", admin: false};
+    return {user};
+};