+++ integrationTests/cypress/support/step_definitions/load-integration-test-libary.ts
@@ -1,0 +1,4 @@
+import doguTestLibrary from "@cloudogu/dogu-integration-test-library";
+
+// Loads all steps from the dogu e2e library into this project
+doguTestLibrary.registerSteps();