+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.10.0-1] - 2023-04-20
+### Changed
+- Rewrite frontend with React (#77)
+- Extend search endpoint to pass a list of excluded values (#77)
+- Update documentation with current screenshots and better explanations (#81)
+- Update cypress-version for integration-tests (#78)
+
+### Added
+- New integration-tests for rewritten frontend (#78)
+
 ## [v1.9.0-1] - 2022-11-14
 ### Changed
 - Add Backend Endpoint for the User Import via CSV-File (#69)